# spring-server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Student Management System - OpenAPI 3.0
  version: 1.0.11
servers:
- url: /api/v3
...
paths:
...
  /course:
    get:
      description: Returns list of courses
      operationId: get_course_list
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Course'
                type: array
                x-content-type: application/json
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Course not found
      summary: Get list of all courses
      tags:
      - course
      x-accepts: application/json
      x-tags:
      - tag: course
      x-openapi-router-controller: swagger_server.controllers.course_controller
    post:
      description: Add a new course
      operationId: add_course
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        description: Create a new course
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: Successful operation
        "405":
          description: Invalid input
      summary: Add a new course
      tags:
      - course
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: course
      x-openapi-router-controller: swagger_server.controllers.course_controller
  /course/{courseId}:
    delete:
      description: delete a course
      operationId: delete_course
      parameters:
      - description: Course id to delete
        explode: false
        in: path
        name: courseId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: successful operation
        "400":
          description: Invalid course value
      summary: Deletes a course
      tags:
      - course
      x-accepts: application/json
      x-tags:
      - tag: course
      x-openapi-router-controller: swagger_server.controllers.course_controller
    get:
      description: Returns a single course
      operationId: get_course
      parameters:
      - description: ID of course to return
        explode: false
        in: path
        name: courseId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Course not found
      summary: Find course by ID
      tags:
      - course
      x-accepts: application/json
      x-tags:
      - tag: course
      x-openapi-router-controller: swagger_server.controllers.course_controller
    put:
      description: Update an existing course by Id
      operationId: update_course
      parameters:
      - description: Course id to delete
        explode: false
        in: path
        name: courseId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        description: Update an existent course
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Course not found
        "405":
          description: Validation exception
      summary: Update an existing course
      tags:
      - course
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: course
      x-openapi-router-controller: swagger_server.controllers.course_controller
components:
...
  schemas:
...
    Course:
      example:
        name: name
        id: 1
        desc: desc
      properties:
        id:
          example: 1
          format: int64
          type: integer
        name:
          type: string
        desc:
          type: string
      type: object
      xml:
        name: course